複合代入の実装方法変更 / virtualな関数は本文なしで定義可能に(ただしそのまま呼び出すとエラーになる)
